# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kabadiwala Connect API",
    "description": "Smart India Hackathon 2026 (PS 26229) - Informal to Formal E-Waste Bridge API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/handovers/initiate": {
      "post": {
        "tags": [
          "Handovers"
        ],
        "summary": "Initiate Handover",
        "operationId": "initiate_handover_api_v1_handovers_initiate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateHandoverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverTransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/handovers/verify": {
      "post": {
        "tags": [
          "Handovers"
        ],
        "summary": "Verify Handover",
        "operationId": "verify_handover_api_v1_handovers_verify_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyHandoverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverTransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/handovers/stream/{lot_id}": {
      "get": {
        "tags": [
          "Handovers"
        ],
        "summary": "Stream Lot Events",
        "operationId": "stream_lot_events_api_v1_handovers_stream__lot_id__get",
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/collector": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Get Collector Ledger",
        "operationId": "get_collector_ledger_api_v1_ledger_collector_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/recycler/dashboard/stats": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Get Recycler Stats",
        "operationId": "get_recycler_stats_api_v1_ledger_recycler_dashboard_stats_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandoverTransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "lot_id": {
            "type": "string",
            "title": "Lot Id"
          },
          "collector_id": {
            "type": "string",
            "title": "Collector Id"
          },
          "recycler_id": {
            "type": "string",
            "title": "Recycler Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "total_payout": {
            "type": "number",
            "title": "Total Payout"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "qr_token": {
            "type": "string",
            "title": "Qr Token"
          },
          "fallback_code": {
            "type": "string",
            "title": "Fallback Code"
          },
          "tamper_check_passed": {
            "type": "boolean",
            "title": "Tamper Check Passed"
          },
          "scanned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "collector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Name"
          },
          "recycler_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lot_id",
          "collector_id",
          "recycler_id",
          "category",
          "weight_kg",
          "unit_price",
          "total_payout",
          "status",
          "qr_token",
          "fallback_code",
          "tamper_check_passed",
          "created_at"
        ],
        "title": "HandoverTransactionResponse"
      },
      "InitiateHandoverRequest": {
        "properties": {
          "lot_id": {
            "type": "string",
            "title": "Lot Id"
          },
          "recycler_id": {
            "type": "string",
            "title": "Recycler Id"
          }
        },
        "type": "object",
        "required": [
          "lot_id",
          "recycler_id"
        ],
        "title": "InitiateHandoverRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyHandoverRequest": {
        "properties": {
          "qr_token": {
            "type": "string",
            "title": "Qr Token"
          },
          "fallback_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Code"
          },
          "actual_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Weight Kg"
          },
          "actual_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Condition"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "qr_token"
        ],
        "title": "VerifyHandoverRequest"
      }
    }
  }
}